share & save visits to calendar

+++ mch_client/src/main/java/com/mch/registry/ccs/data/adapter/RecommendationArrayAdapter.java
@@ -20,7 +20,7 @@ public class RecommendationArrayAdapter extends TwoLineArrayAdapter<Recommendati
 
 	@Override
 	public String lineOneText(Recommendation r) {
-		return _context.getString(R.string.pregnancy_week) +  ": " + Double.toString(Math.floor(r.get_recommendationDay()/7));
+		return _context.getString(R.string.pregnancy_week) +  ": " + Integer.toString(((int) Math.floor(r.get_recommendationDay() / 7)));
 	}
 
 	@Override
